Only reset src/ngeo if src/ngeo is initialized as a subrepo
When we clone the repo, the src/ngeo folder exists but is empty. This means that the first 'manuel update' will reset the geo-front3 repo. We test that src/ngeo is initialized before resetting it to prevent this error.

+++ tasks/utils.sh
@@ -683,7 +683,7 @@ function update-ngeo {
     local ngeo_patches_dir="$(realpath ${FRONT_DIR}/scripts/ngeo-patches)"
 
     pushd "${FRONT_DIR}"
-        if [[ -d src/ngeo ]]; then
+        if [[ -d src/ngeo && -f src/ngeo/.git ]]; then
             pushd src/ngeo
                 # Unapply all patches.
                 git reset --hard
